gilbenizer/requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.2
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
decorator==4.2.1
dill==0.2.7.1
entrypoints==0.2.3
enum34==1.1.6
ffmpeg-python==0.1.10
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gevent==1.2.2
google-api-core==0.1.4
google-auth==1.3.0
google-cloud-core==0.28.0
google-cloud-speech==0.30.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
greenlet==0.4.12
grpcio==1.8.4
html5lib==1.0.1
idna==2.6
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.3.1
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
ply==3.8
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
scandir==1.6
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.2
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
configparser==3.5.0
decorator==4.2.1
dill==0.2.7.1
entrypoints==0.2.3
enum34==1.1.6
ffmpeg-python==0.1.10
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gevent==1.2.2
google-api-core==0.1.4
google-auth==1.3.0
google-cloud-core==0.28.0
google-cloud-speech==0.30.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
greenlet==0.4.12
grpcio==1.8.4
html5lib==1.0.1
idna==2.6
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.3.1
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
ply==3.8
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
rsa==3.4.2
scandir==1.6
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
